organize string funcs and add other utility str funcs
